calculate text size based on video size and screen size
