+++ app/src/main/java/app/myzel394/alibi/effect/opengl/ShaderSource.kt
@@ -1,4 +1,4 @@
-package app.myzel394.alibi.services.effect.opengl
+package app.myzel394.alibi.effect.opengl
 
 /**
  * Represents the shader source strings.
